fix(04): fix invisible tint color, deduplicate food auto-caching, mock api for testing
test(04): complete UAT - 4 passed, 2 issues resolved

+++ LiiO_EatClean/Features/Profile/ProfileViewModel.swift
@@ -56,7 +56,7 @@ class ProfileViewModel {
         let updatedUser = UserModel(
             id: user?.id ?? UUID(),
             name: name,
-            age: Int16(age) ?? user?.age ?? 0,
+            age: Double(age) ?? user?.age ?? 0,
             gender: user?.gender ?? "male",
             height: Double(height) ?? user?.height ?? 0,
             weight: Double(weight) ?? user?.weight ?? 0,
